Allow selecting Linux payload default DTB (#3)

+++ scripts/build-firmware-linux.sh
@@ -7,6 +7,7 @@ DTS_TEMPLATE_DIR="$(realpath "$3")"
 KERNEL_IMAGE="$(realpath "$4")"
 WORKLOAD_BUILD_DIR="$(realpath "$5")"
 CPIO_ARCHIVE="$WORKLOAD_BUILD_DIR/rootfs.cpio"
+DEFAULT_DTB="${DEFAULT_DTB:-xiangshan}"
 
 MEM_BEGIN=$(( 0x80000000 ))
 DTB_OFFSET_KB=1536
@@ -40,9 +41,14 @@ for dts_template in "$DTS_TEMPLATE_DIR"/*.dts.in ; do
 done
 
 # Assemble the image
-# Using `xiangshan.dtb` as the "defualt" device tree
+# Using `xiangshan.dtb` as the default device tree unless DEFAULT_DTB is set.
+DEFAULT_DTB_FILE="$WORKLOAD_BUILD_DIR/dt/$DEFAULT_DTB.dtb"
+if ! [ -f "$DEFAULT_DTB_FILE" ]; then
+    echo "Default device tree not found: $DEFAULT_DTB_FILE" >&2
+    exit 1
+fi
 dd if="$STARTUP_FILE" of="$WORKLOAD_BUILD_DIR/fw_payload.bin"
-dd if="$WORKLOAD_BUILD_DIR/dt/xiangshan.dtb" of="$WORKLOAD_BUILD_DIR/fw_payload.bin" bs="$KILOBYTE" seek="$DTB_OFFSET_KB" conv=notrunc
+dd if="$DEFAULT_DTB_FILE" of="$WORKLOAD_BUILD_DIR/fw_payload.bin" bs="$KILOBYTE" seek="$DTB_OFFSET_KB" conv=notrunc
 dd if="$SBI_BUILD_DIR/build/platform/generic/firmware/fw_jump.bin" of="$WORKLOAD_BUILD_DIR/fw_payload.bin" bs="$KILOBYTE" seek="$SBI_OFFSET_KB" conv=notrunc
 dd if="$KERNEL_IMAGE" of="$WORKLOAD_BUILD_DIR/fw_payload.bin" bs="$MEGABYTE" seek="$KERNEL_OFFSET_MB" conv=notrunc
 dd if="$CPIO_ARCHIVE" of="$WORKLOAD_BUILD_DIR/fw_payload.bin" bs="$MEGABYTE" seek="$INITRAMFS_OFFSET_MB" conv=notrunc
